// apps/fleet/fleet-mfe/src/app/fleet-product/line/shared/models/sales-order/line-item.model.ts
import { OrderItem } from './order-item.model';

export class LineItem {
  orderItems: OrderItem[];
  productIdentifier: string;
  publicateNumber: boolean;
  noProductChanges?: boolean;
}

export enum ProductTypeEnum {
  MOBILE = 'MOBILE',
  OPTION = 'OPTION',
  SMARTPHONE = 'SMARTPHONE',
  SMARTPHONES = 'SMARTPHONES',
  SMARTPHONE_OTC = 'SMARTPHONE OTC',
  CANCELLATION_FEE = 'CANCELLATION FEE',
  CONTENT_DISCONNECTION_FEE = 'CONTENT DISCONNECTION FEE',
  PORT_IN_PHONE_NUMBER = 'PORT-IN PHONE NUMBER',
  SIMCARD = 'SIM CARD',
  INTERNET = 'INTERNET',
  DTV = 'DTV',
  TV = 'OTTV',
  FIXED_LINE = 'FIXED LINE',
  BUNDLE = 'BUNDLE',
  PLAN = 'PLAN',
  TELEPHONE = 'TELEPHONE',
  TRIAL_PACK = 'TRIALPACK',
  MOBILE_LINE = 'MOBILE LINE',
  IDTV_LINE = 'IDTV LINE',
  DTV_LINE = 'DTV LINE',
  TV_BOX = 'TV_BOX',
  INTERNET_LINE = 'INTERNET LINE',
  HARDWARE = 'HARDWARE',
  OTT_LINE = 'OTT LINE',
  DATA_ONLY_MOBILE_LINE = 'DATA ONLY MOBILE LINE',
  PLACEHOLDER = 'PLACEHOLDER',
  HARDWARE_SMARTHOME_IDENTIFIER = 'SMARTHOME',
  HARDWARE_ACCESSORIES_IDENTIFIER = 'ACCESSORIES',
  HARDWARE_HOME_OFFICE_IDENTIFIER = 'HOME_OFFICE',
  HARDWARE_SMARTWATCH_IDENTIFIER = 'SMARTWATCH',
  HARDWARE_TV_AUDIO_IDENTIFIER = 'TV_AUDIO',
  HARDWARE_TABLET_IDENTIFIER = 'TABLET',
  HARDWARE_MULTIMEDIA_IDENTIFIER = 'MULTIMEDIA',
  HARDWARE_PROMO = 'PROMO',
  HARDWARE_SMARTHOME_SOHO_IDENTIFIER = 'SMARTHOME_SOHO',
  HARDWARE_ACCESSORIES_SOHO_IDENTIFIER = 'ACCESSORIES_SOHO',
  HARDWARE_HOME_OFFICE_SOHO_IDENTIFIER = 'HOME_OFFICE_SOHO',
  HARDWARE_SMARTWATCH_SOHO_IDENTIFIER = 'SMARTWATCH_SOHO',
  HARDWARE_TV_AUDIO_SOHO_IDENTIFIER = 'TV_AUDIO_SOHO',
  HARDWARE_PROMO_SOHO = 'PROMO_SOHO',
  HARDWARE_MULTIMEDIA_SOHO_IDENTIFIER = 'MULTIMEDIA_SOHO',
  HARDWARE_TABLET_SOHO_IDENTIFIER = 'TABLET_SOHO',
  HARDWARE_SMARTPHONE_SOHO_IDENTIFIER = 'SMARTPHONE_SOHO',
  TV_EASY_SWITCH = 'TV',
  EXPERIENCE = 'EXPERIENCE',
  DOMAIN_NAME = 'DOMAIN NAME',
  PLUME = 'PLUME',
  YEAR_END_PROMO = 'PROMO_EOY',
  CURRENT_PROVIDER_CLICK_EVENT_NAME = 'Current provider option clicked',
  CURRENT_PROVIDER_SELECTED_EVENT_NAME = 'Current provider selected',
  // eslint-disable-next-line @typescript-eslint/no-duplicate-enum-values
  INTERNET_EASY_SWITCH = 'INTERNET',
  SPRING_DEAL = 'SPRING_DEAL',
  HARDWARE_OUTLET_IDENTIFIER = 'OUTLET',
  HARDWARE_OUTLET_SOHO_IDENTIFIER = 'OUTLET_SOHO',
  HARDWARE_ONE_DEVICE_PROMO_IDENTIFIER = 'SUMMER_DEALS',
  HARDWARE_ONE_DEVICE_PROMO_SOHO_IDENTIFIER = 'SUMMERPROMO_SOHO',
  HARDWARE_PROMO_FEEST_PROMO = 'PROMO_FEEST_PROMO',
  HARDWARE_PROMO_FEEST_PROMO_SOHO = 'PROMO_FEEST_PROMO_SOHO',
  HARDWARE_ECO_FRIENDLY = 'ECO_FRIENDLY',
  HARDWARE_ECO_FRIENDLY_SOHO = 'ECO_FRIENDLY_SOHO',
  GAMING = 'GAMING',
  GAMING_SOHO = 'GAMING_SOHO',
  PHONE_TELENET = 'PHONE_TELENET',
  PHONE_TELENET_SOHO = 'PHONE_TELENET_SOHO',
}

export enum ProductStatusEnum {
  ADD = 'ADD',
  ADDED = 'ADDED',
  MODIFY = 'MODIFY',
  ACTIVE = 'ACTIVE',
  NOT_ACTIVE = 'NOT_ACTIVE',
  PENDINGACTIVATION = 'PENDING_FOR_ACTIVATION',
  ACTIVATION_PASSED_PONR = 'ACTIVATION_PASSED_PONR',
  DISCONNECT = 'DISCONNECT',
  DISCONNECTED = 'DISCONNECTED',
  COMPLETED = 'COMPLETED',
  SUSPENDED = 'SUSPENDED',
  DISCONNECTION_IN_PROGRESS = 'DISCONNECTION_IN_PROGRESS',
}